controllers update

+++ backend/controllers/user.controllers.js
@@ -4,9 +4,10 @@ const db = require("../models");
 const User = db.User;
 const sequelize = db.sequelize;
 const Op = db.Sequelize.Op;
+const uploadFiles = require("../utils/multer.utils");
 const transporter = require("../utils/nodemailer.utils");
 const generateOTP = require("../utils/generateOTP.utils");
-const imagekit = require("../utils/imageKit.utils");
+const { imagekit, sendToImageKit } = require("../utils/imageKit.utils");
 
 // @desc Auth user & get token
 // @route POST /api/users/signin
@@ -306,32 +307,49 @@ const resetPassword = asyncHandler(async (req, res) => {
 // @desc update user profile picture
 // @route POST /user/profile-picture
 // @access Private
-const updateDP = asyncHandler(async (req, res) => {
-  console.log({ body: req.body });
-  const user = await User.findByPk(req.params.id, {
-    attributes: { exclude: ["password"] },
+const updateDP = asyncHandler(async (req, res, next) => {
+  uploadFiles(req, res, async (err) => {
+    try {
+      if (err instanceof multer.MulterError) {
+        res.status(400);
+        throw new Error(err.message);
+      } else if (err) {
+        res.status(400);
+        throw new Error(err.toString());
+      } else {
+        const user = await User.findByPk(req.params.id, {
+          attributes: { exclude: ["password"] },
+        });
+        const defaultUrl =
+          "https://ik.imagekit.io/msf9dwhbk3m/AC26972E-41EB-4318-877F-1ACBB7B32AC0_IMrRiKVMPG-t.jpeg";
+        if (user) {
+          const { image } = req.files;
+          let update;
+          if (image) update = await sendToImageKit(image[0]);
+          else {
+            update = {
+              url: defaultUrl,
+            };
+          }
+          if (
+            user.profileImage.url !== defaultUrl &&
+            user.profileImage.url !== update.url
+          ) {
+            await imagekit.deleteFile(user.profileImage.fileId);
+          }
+          await user.update({ profileImage: update });
+          res
+            .status(200)
+            .json({ ...user.dataValues, token: generateToken(user.id) });
+        } else {
+          res.status(404);
+          throw new Error("User not found");
+        }
+      }
+    } catch (err) {
+      next(err);
+    }
   });
-  const defaultUrl =
-    "https://ik.imagekit.io/msf9dwhbk3m/AC26972E-41EB-4318-877F-1ACBB7B32AC0_IMrRiKVMPG-t.jpeg";
-  if (user) {
-    let update;
-    const dp = req.body.dp;
-    if (dp === {}) {
-      update = {
-        url: defaultUrl,
-      };
-    } else {
-      update = dp;
-    }
-    if (user.profileImage.url !== defaultUrl) {
-      await imagekit.deleteFile(user.profileImage.fileId);
-    }
-    await user.update({ profileImage: update });
-    res.status(200).json({ ...user.dataValues, token: generateToken(user.id) });
-  } else {
-    res.status(404);
-    throw new Error("User not found");
-  }
 });
 
 module.exports = {
